Create config package and fix dependencies
