+++ packages/0xsequence/tests/browser/proxy-transport/channel.test.ts
@@ -3,7 +3,8 @@ import { ethers, Wallet as EOAWallet } from 'ethers'
 import { Web3Provider, JsonRpcProvider } from '@ethersproject/providers'
 import { test, assert } from '../../utils/assert'
 import { sequenceContext, testnetNetworks } from '@0xsequence/network'
-import { Wallet, addressOf, isValidSignature, packMessageData, recoverConfig } from '@0xsequence/wallet'
+import { Wallet, isValidSignature, packMessageData, recoverConfig } from '@0xsequence/wallet'
+import { addressOf } from '@0xsequence/config'
 import { LocalRelayer } from '@0xsequence/relayer'
 import { testAccounts, getEOAWallet } from '../testutils'
