+++ packages/provider/package.json
@@ -16,8 +16,11 @@
     "@0xsequence/auth": "^0.7.0",
     "@0xsequence/network": "^0.7.0",
     "@0xsequence/transactions": "^0.7.0",
+    "@0xsequence/utils": "^0.7.0",
     "@0xsequence/wallet": "^0.7.0",
+    "@ethersproject/hash": "^5.0.10",
     "@ethersproject/providers": "^5.0.17",
+    "@ethersproject/web": "^5.0.12",
     "ethers": "^5.0.24",
     "ethers-eip712": "^0.2.0",
     "eventemitter3": "^4.0.7",
