+++ packages/provider/src/provider.ts
@@ -5,7 +5,8 @@ import { TypedDataDomain, TypedDataField, TypedDataSigner } from '@ethersproject
 import { _TypedDataEncoder } from '@ethersproject/hash'
 import { poll } from '@ethersproject/web'
 import { NetworkConfig, Networks, WalletContext, ChainId, JsonRpcHandler, JsonRpcHandlerFunc, JsonRpcFetchFunc, JsonRpcRequest, JsonRpcResponseCallback, JsonRpcResponse, maybeNetworkId, JsonRpcVersion } from '@0xsequence/network'
-import { Signer, WalletConfig, WalletState } from '@0xsequence/wallet'
+import { Signer } from '@0xsequence/wallet'
+import { WalletConfig, WalletState } from '@0xsequence/config'
 import { Relayer } from '@0xsequence/relayer'
 import { Deferrable, shallowCopy, resolveProperties } from '@0xsequence/utils'
 import { TransactionRequest, TransactionResponse, Transactionish, SignedTransactions } from '@0xsequence/transactions'
