+++ packages/wallet/package.json
@@ -16,11 +16,15 @@
   "dependencies": {
     "@0xsequence/abi": "^0.7.0",
     "@0xsequence/api": "^0.7.0",
+    "@0xsequence/config": "^0.7.0",
     "@0xsequence/guard": "^0.7.0",
     "@0xsequence/network": "^0.7.0",
+    "@0xsequence/relayer": "^0.7.0",
     "@0xsequence/transactions": "^0.7.0",
     "@0xsequence/utils": "^0.7.0",
     "@arcadeum/ethauth": "^0.3.0",
+    "@ethersproject/abi": "^5.0.10",
+    "@ethersproject/properties": "^5.0.7",
     "@ethersproject/providers": "^5.0.17",
     "ethers": "^5.0.24",
     "ethers-eip712": "^0.2.0",
