+++ packages/wallet/src/account.ts
@@ -5,7 +5,7 @@ import { Deferrable } from '@ethersproject/properties'
 import { walletContracts } from '@0xsequence/abi'
 import { Signer, NotEnoughSigners } from './signer'
 import { SignedTransactions, Transactionish } from '@0xsequence/transactions'
-import { WalletConfig, WalletState, addressOf, imageHash, isConfigEqual } from './config'
+import { WalletConfig, WalletState, addressOf, imageHash, isConfigEqual } from '@0xsequence/config'
 import { ChainId, NetworkConfig, WalletContext, sequenceContext, mainnetNetworks, isNetworkConfig, ensureValidNetworks, sortNetworks, getNetworkId } from '@0xsequence/network'
 import { Wallet } from './wallet'
 import { resolveArrayProperties, findLatestLog } from './utils'
